feat/manufacturer_item_history

+++ src/main/java/com/springboot/item_manufacture/dto/Dto.java
@@ -6,6 +6,7 @@ import javax.validation.constraints.DecimalMin;
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotBlank;
 import java.math.BigDecimal;
+import java.time.LocalDateTime;
 
 public class Dto {
     @Getter
@@ -22,7 +23,8 @@ public class Dto {
         private BigDecimal unitPrice;
 
         @Min(value = 1, message = "수량은 1개 이상이어야 합니다.")
-        private int qty;
+        private Long qty;
+
     }
 
     @Getter
@@ -38,6 +40,6 @@ public class Dto {
         private String itemCd; // 아이템 코드
         private String itemNm; // 아이템 명
         private BigDecimal unitPrice; // 제조사 맞춤 단가
-        private int qty;        // 제품 수량
+        private Long qty;        // 제품 수량
     }
 }
